add budget to optimizer

+++ shoppingOptimization/src/main/java/Item.java
@@ -1,6 +1,6 @@
 public class Item {
     public String itemName;
     public double price;
-    public int qualityPoint;
+    public double qualityPoint;
     public ItemGroup group;
 }
